french-writing-practice/requirements.txt
streamlit==1.31.0
pillow==10.2.0
opencv-python==4.9.0.80
pytesseract==0.3.10
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
python-i18n==0.3.9
scikit-image==0.22.0